+++ attendance-api/src/database/visitedUser/fetch.ts
@@ -1,7 +1,7 @@
 import { ModelStatic } from "sequelize";
-import { VisitedUserModel } from "../../types/visitedUser";
+import { ModifiedVisitedUserModel, VisitedUserModel } from "../../types/visitedUser";
 
-export async function fetch(VisitedUser: ModelStatic<any>, filter?: Partial<VisitedUserModel>): Promise<VisitedUserModel[]> {
+export async function fetch(VisitedUser: ModelStatic<any>, filter?: Partial<VisitedUserModel>): Promise<ModifiedVisitedUserModel[]> {
 	let options: object = filter ? {
 		where: filter
 	} : {}